1.1.0+2025/07/01
Se realizan DataLoader, Comentarios con Swagger y se empieza el desarrollo de las pruebas unitarias, falta su revisión y ajustes para adaptarse a los criterios necesarios

+++ src/main/java/com/SAFE_Rescue/API_Turno/repository/BomberoRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
  * y acceso a datos utilizando Spring Data JPA.
  */
 @Repository
-public interface BomberoRepository extends JpaRepository<Bombero, Long>{
+public interface BomberoRepository extends JpaRepository<Bombero, Integer>{
 
 
 }
