+++ src/main/java/com/SAFE_Rescue/API_Turno/repository/TurnoRepository.java
@@ -12,6 +12,6 @@ import java.util.List;
  * y acceso a datos utilizando Spring Data JPA.
  */
 @Repository
-public interface TurnoRepository extends JpaRepository<Turno, Long> {
+public interface TurnoRepository extends JpaRepository<Turno, Integer> {
 
 }
